+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.22 (2021-02-24)
+
+### Bug Fixes
+
+- [x] la descrizione non rispetta gli "a capo" inseriti nel testo
+
+### Features
+
+- [x] iscrizione a singoli seminari
+
 ## 1.0.21 (2021-02-22)
 
 ### Bug Fixes